+++ crates/bcr-wallet-core/src/wallet/mod.rs
@@ -1,7 +1,11 @@
 mod api;
 mod builder;
+mod credit;
+mod types;
 mod utils;
 mod wallet;
 
 pub use builder::*;
+pub use credit::*;
+pub use types::*;
 pub use wallet::*;
