+++ crates/bcr-wallet-core/src/wallet/types.rs
@@ -1,0 +1,10 @@
+use anyhow::Result;
+use cashu::{Amount, Proof};
+
+pub trait SwapProofs {
+    async fn swap_proofs_amount(
+        &self,
+        proofs: Vec<Proof>,
+        amounts: Vec<Amount>,
+    ) -> Result<Vec<Proof>>;
+}
